Development

See merge request ds2021_30442_petrean_florin/ds2021_30442_petrean_florin_1_backend_assig2!3

+++ src/main/java/com/ds/EnergyUtilityPlatform/controller/UserController.java
@@ -1,6 +1,8 @@
 package com.ds.EnergyUtilityPlatform.controller;
 
 import com.ds.EnergyUtilityPlatform.model.dto.*;
+import com.ds.EnergyUtilityPlatform.model.dto.entitydto.DeviceDto;
+import com.ds.EnergyUtilityPlatform.model.dto.entitydto.UserDto;
 import com.ds.EnergyUtilityPlatform.model.entity.AppUser;
 import com.ds.EnergyUtilityPlatform.model.entity.Device;
 import com.ds.EnergyUtilityPlatform.service.ICrudService;
@@ -36,7 +38,7 @@ public class UserController extends CrudController<AppUser, UserDto> {
     }
 
     @GetMapping(path = "/{username}/totalEnergy")
-    public Float getUserTotalEnergy(@PathVariable String username) {
+    public Double getUserTotalEnergy(@PathVariable String username) {
         return userService.getEnergyConsumption(username);
     }
